requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
langchain>=0.3.0
langchain-openai>=0.1.1
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-core>=0.2.0
chromadb>=0.4.22
pydantic>=2.5.3
python-dotenv>=1.0.0
python-multipart>=0.0.9
openai>=1.12.0
tiktoken>=0.6.0
pytesseract>=0.3.10
pdf2image>=1.16.3
unstructured>=0.10.30
jinja2==3.1.2
